# lib/api-zod/src/generated/types/activityDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from "./activityEvent";
import type { GeoPoint } from "./geoPoint";

export type ActivityDetail = ActivityEvent & {
  location?: GeoPoint | null;
  consumerNo?: string | null;
  reading?: number | null;
  photoUri?: string | null;
  selfieUri?: string | null;
  notes?: string | null;
  distanceKm?: number | null;
  durationSec?: number | null;
  origin?: GeoPoint | null;
  destination?: GeoPoint | null;
};