# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityKind } from "./activityKind";

export interface ActivityEvent {
  id: string;
  kind: ActivityKind;
  staffId: string;
  staffName: string;
  occurredAt: Date;
  receivedAt: Date;
  synced: boolean;
  tripRef?: string | null;
  /** Pre-formatted human-readable summary for list views */
  summary: string;
}