# lib/api-zod/src/generated/types/activityKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ActivityKind = (typeof ActivityKind)[keyof typeof ActivityKind];

export const ActivityKind = {
  checkin: "checkin",
  checkout: "checkout",
  meter: "meter",
  "trip-start": "trip-start",
  "trip-end": "trip-end",
} as const;