# lib/api-zod/src/generated/types/activityPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from "./activityEvent";

export interface ActivityPage {
  items: ActivityEvent[];
  nextCursor: string | null;
  /** Server clock at response — clients should send this back as `since` on the next poll. */
  serverTime: Date;
}