# lib/api-zod/src/generated/types/createActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityKind } from "./activityKind";
import type { GeoPoint } from "./geoPoint";

export interface CreateActivityInput {
  kind: ActivityKind;
  staffId: string;
  staffName: string;
  occurredAt?: Date;
  tripRef?: string | null;
  location?: GeoPoint | null;
  consumerNo?: string | null;
  reading?: number | null;
  photoUri?: string | null;
  selfieUri?: string | null;
  notes?: string | null;
  distanceKm?: number | null;
  durationSec?: number | null;
  origin?: GeoPoint | null;
  destination?: GeoPoint | null;
}