# lib/api-zod/src/generated/types/staff.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StaffRole } from "./staffRole";

export interface Staff {
  id: string;
  empCode: string;
  name: string;
  phone: string;
  role: StaffRole;
}